requirements.txt
-e .

aws-cdk.core
aws-cdk.aws_lambda
aws_cdk.aws_apigatewayv2
aws_cdk.aws_apigatewayv2_integrations